Changed "lang-import", so that ONLY language dependent sites have to be in folder "/_lang/xy/".
I have changed the behaviour of the templators (and layouts) language
feature.
Formerly, any page had to be in "/_lang/xy/". New, if a Page is NOT
language depandent (except property entries), it has not to be
repeated in each languag.
The page includes itselfs, but the layout is only included once (include_once dir.php).